rename header files to .h and add license
