+++ tools/protoc-gen-chpl/protoc-gen-chpl.cpp
@@ -20,7 +20,7 @@
 
 #include <google/protobuf/compiler/plugin.h>
 
-#include <generator.hh>
+#include <generator.h>
 
 int main(int argc, char *argv[]) {
   chapel::Generator generator;
