+++ tools/protoc-gen-chpl/repeated_enum_field.h
@@ -18,19 +18,27 @@
  * limitations under the License.
  */
 
-#include <primitive_field.h>
+#ifndef PB_REPEATED_ENUM_FIELD_HH
+#define PB_REPEATED_ENUM_FIELD_HH
+
+#include <google/protobuf/io/printer.h>
+#include <google/protobuf/descriptor.h>
+
+#include <field_base.h>
 
 namespace chapel {
+  
+  using namespace google::protobuf;
+  using namespace google::protobuf::io;
+  
+  class RepeatedEnumFieldGenerator : public FieldGeneratorBase {
+   public:
+    RepeatedEnumFieldGenerator(const FieldDescriptor* descriptor);
+    ~RepeatedEnumFieldGenerator();
 
-  PrimitiveFieldGenerator::PrimitiveFieldGenerator(const FieldDescriptor* descriptor)
-      : FieldGeneratorBase(descriptor) {
-  }
-
-  PrimitiveFieldGenerator::~PrimitiveFieldGenerator() {
-  }
-
-  void PrimitiveFieldGenerator::GenerateMembers(Printer* printer) {
-    printer->Print(variables_, "var $name$: $type_name$;\n");
-  }
+    void GenerateMembers(Printer* printer);
+  };
 
 }  // namespace chapel
+
+#endif /* PB_REPEATED_ENUM_FIELD_HH */
