+++ tools/protoc-gen-chpl/repeated_primitive_field.h
@@ -18,19 +18,27 @@
  * limitations under the License.
  */
 
-#include <primitive_field.h>
+#ifndef PB_REPEATED_PRIMITIVE_FIELD_HH
+#define PB_REPEATED_PRIMITIVE_FIELD_HH
+
+#include <google/protobuf/io/printer.h>
+#include <google/protobuf/descriptor.h>
+
+#include <field_base.h>
 
 namespace chapel {
+  
+  using namespace google::protobuf;
+  using namespace google::protobuf::io;
+  
+  class RepeatedPrimitiveFieldGenerator : public FieldGeneratorBase {
+   public:
+    RepeatedPrimitiveFieldGenerator(const FieldDescriptor* descriptor);
+    ~RepeatedPrimitiveFieldGenerator();
 
-  PrimitiveFieldGenerator::PrimitiveFieldGenerator(const FieldDescriptor* descriptor)
-      : FieldGeneratorBase(descriptor) {
-  }
-
-  PrimitiveFieldGenerator::~PrimitiveFieldGenerator() {
-  }
-
-  void PrimitiveFieldGenerator::GenerateMembers(Printer* printer) {
-    printer->Print(variables_, "var $name$: $type_name$;\n");
-  }
+    void GenerateMembers(Printer* printer);
+  };
 
 }  // namespace chapel
+
+#endif /* PB_REPEATED_PRIMITIVE_FIELD_HH */
